Extract RunWithManagedStdio from subprocess.h
https://github.com/google/android-cuttlefish/pull/1432/files#r2226344612 points out that many callers have the same logic around calling the interface. As a first step, extracting to a separate file makes it easier to focus on modifying this interface.

Bug: b/433812518
