+++ base/cvd/cuttlefish/common/libs/utils/archive.cpp
@@ -27,6 +27,7 @@
 
 #include "cuttlefish/common/libs/utils/result.h"
 #include "cuttlefish/common/libs/utils/subprocess.h"
+#include "cuttlefish/common/libs/utils/subprocess_managed_stdio.h"
 
 namespace cuttlefish {
 namespace {
