+++ base/cvd/cuttlefish/common/libs/utils/disk_usage.cpp
@@ -24,6 +24,7 @@
 
 #include "cuttlefish/common/libs/utils/result.h"
 #include "cuttlefish/common/libs/utils/subprocess.h"
+#include "cuttlefish/common/libs/utils/subprocess_managed_stdio.h"
 
 namespace cuttlefish {
 namespace {
