+++ base/cvd/cuttlefish/common/libs/utils/wait_for_unix_socket.cpp
@@ -23,6 +23,7 @@
 #include "cuttlefish/common/libs/utils/files.h"
 #include "cuttlefish/common/libs/utils/result.h"
 #include "cuttlefish/common/libs/utils/subprocess.h"
+#include "cuttlefish/common/libs/utils/subprocess_managed_stdio.h"
 
 namespace cuttlefish {
