+++ base/cvd/cuttlefish/host/commands/powerbtn_cvd/BUILD.bazel
@@ -13,7 +13,7 @@ cf_cc_binary(
         "//cuttlefish/common/libs/fs",
         "//cuttlefish/common/libs/utils:result",
         "//cuttlefish/common/libs/utils:subprocess",
-        "//cuttlefish/host/libs/command_util",
+        "//cuttlefish/common/libs/utils:subprocess_managed_stdio",
         "//cuttlefish/host/libs/config:cuttlefish_config",
         "//libbase",
         "@gflags",
