+++ base/cvd/cuttlefish/host/commands/start/flag_forwarder.cc
@@ -27,10 +27,9 @@
 #include <gflags/gflags.h>
 #include <libxml/parser.h>
 
-#include "cuttlefish/common/libs/fs/shared_buf.h"
-#include "cuttlefish/common/libs/fs/shared_fd.h"
 #include "cuttlefish/common/libs/utils/contains.h"
 #include "cuttlefish/common/libs/utils/subprocess.h"
+#include "cuttlefish/common/libs/utils/subprocess_managed_stdio.h"
 
 namespace cuttlefish {
