+++ base/cvd/cuttlefish/host/libs/avb/BUILD.bazel
@@ -17,6 +17,7 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/common/libs/utils:result",
         "//cuttlefish/common/libs/utils:subprocess",
+        "//cuttlefish/common/libs/utils:subprocess_managed_stdio",
         "//cuttlefish/host/libs/config:known_paths",
         "@fruit",
     ],
