+++ base/cvd/cuttlefish/host/libs/avb/avb.cpp
@@ -28,6 +28,7 @@
 #include "cuttlefish/common/libs/utils/files.h"
 #include "cuttlefish/common/libs/utils/result.h"
 #include "cuttlefish/common/libs/utils/subprocess.h"
+#include "cuttlefish/common/libs/utils/subprocess_managed_stdio.h"
 #include "cuttlefish/host/libs/config/known_paths.h"
 
 namespace cuttlefish {
