+++ base/cvd/cuttlefish/host/libs/config/config_utils.cpp
@@ -33,6 +33,7 @@
 #include "cuttlefish/common/libs/utils/files.h"
 #include "cuttlefish/common/libs/utils/in_sandbox.h"
 #include "cuttlefish/common/libs/utils/subprocess.h"
+#include "cuttlefish/common/libs/utils/subprocess_managed_stdio.h"
 #include "cuttlefish/host/libs/config/config_constants.h"
 
 namespace cuttlefish {
