+++ base/cvd/cuttlefish/host/libs/vm_manager/BUILD.bazel
@@ -39,6 +39,7 @@ cf_cc_library(
         "//cuttlefish/common/libs/utils:network",
         "//cuttlefish/common/libs/utils:result",
         "//cuttlefish/common/libs/utils:subprocess",
+        "//cuttlefish/common/libs/utils:subprocess_managed_stdio",
         "//cuttlefish/common/libs/utils:users",
         "//cuttlefish/common/libs/utils:wait_for_unix_socket",
         "//cuttlefish/host/libs/command_util",
