+++ base/cvd/cuttlefish/host/libs/vm_manager/crosvm_display_controller.cpp
@@ -23,6 +23,7 @@
 #include <vector>
 #include "cuttlefish/common/libs/utils/result.h"
 #include "cuttlefish/common/libs/utils/subprocess.h"
+#include "cuttlefish/common/libs/utils/subprocess_managed_stdio.h"
 #include "cuttlefish/host/libs/config/cuttlefish_config.h"
 
 namespace cuttlefish {
